requirements.txt
pandas
yfinance
fredapi
python-dotenv
fear-and-greed
yahoo_fin
tensorflow
scikit-learn
joblib
optuna
# tensorflow-rocm
binance-connector
python-binance
requests
coinglass-api